LayoutTests/imported/w3c/web-platform-tests/speculation-rules/header-only-top-level-expected.txt


FAIL Speculation-Rules header must only be fetched for top-level windows assert_equals: rules must not be fetched for iframes expected 1 but got 2
